[Empty Char Driver] Create a empty char driver example
In this example, we can observe what will happen after the driver module
is loaded.

The driver module will create /sys/class/empty_char_driver.

And we can see there're dev and uevent under the device.
/sys/class/empty_char_driver/ecd/dev
/sys/class/empty_char_driver/ecd/uevent

The dev stores the device number, and uevent stores the corresponding
information for udev to create the device node.

Also, the device node will be created automatically by the VFS.
/dev/ecd